--- android/documentation/changelog.md
@@ -1,14 +1,0 @@
-# Change Log
-<pre>
-v2.0.2  [MOD-2167] Recompiled binary for Android 6.0 support
-
-v2.0.2	Building with 2.1.3.GA to support x86 devices
-	
-v2.0.1	Fixed a couple multi-touch scenarios that were drawing erratic lines [MOD-638]
-
-v2.0	Upgraded to module api version 2 for 1.8.0.1
-
-v1.1	Added multi-touch support [MOD-243]
-		Added "image" property to the paint view. See example and documentation to find out more.
-
-v1.0    Initial Release
