--- android/documentation/index.md
@@ -1,30 +1,0 @@
-# Ti.Paint Module
-
-## Description
-
-Provides a paint surface user interface view.
-
-## Accessing the Ti.Paint Module
-
-To access this module from JavaScript, you would do the following:
-
-	var Paint = require('ti.paint');
-
-### Ti.Paint.createPaintView({...})
-
-Creates a [Ti.Paint.PaintView][] object.
-
-#### Arguments
-
-* parameters [object] - (optional) a dictionary object properties defined in [Titanium.UI.View][]
-                                                                                                 
-## Module History
-
-View the [change log](changelog.html) for this module.
-
-## Usage
-
-See example
-
-[Ti.Paint.PaintView]: paintView.html
-[Titanium.UI.View]: http://developer.appcelerator.com/apidoc/mobile/latest/Titanium.UI.View-object.html
