--- android/documentation/paintView.md
@@ -1,33 +1,0 @@
-# Ti.Paint.PaintView
-
-## Description
-
-A _Ti.Paint_ module object which is a view for painting in.
-
-## Functions
-
-### clear()
-
-Clears the paint view.
-
-## Properties
-
-### strokeWidth[double]
-
-Controls the width of the strokes.
-
-### strokeColor[string]
-
-Controls the color of the strokes.
-
-### strokeAlpha[int]
-
-Controls the opacity of the strokes.
-
-### eraseMode[boolean]
-
-Controls if the strokes are in "erase mode" -- that is, any existing paint will be erased.
-
-### image[string]
-
-Loads an image (by its URL) directly in to the paint view so that it can be drawn on and erased.
