+++ windows/changelog.md
@@ -1,4 +1,4 @@
-# Change Log
+# Change Log (iOS)
 <pre>
 v1.4.1 	Added support for touchstart, touchend, touchcancel, touchmove [MOD-2070]
 
@@ -13,6 +13,22 @@ v1.3	Rewrote rendering algorithm to allow for a number of enhancements:
 
 v1.2	Fixed percent width visual defects [MOD-348]
 		Improved multi-touch drawing performance
+
+v1.1	Added multi-touch support [MOD-243]
+		Added "image" property to the paint view. See example and documentation to find out more.
+
+v1.0    Initial Release
+
+
+# Change Log (android)
+<pre>
+v2.0.2  [MOD-2167] Recompiled binary for Android 6.0 support
+
+v2.0.2	Building with 2.1.3.GA to support x86 devices
+
+v2.0.1	Fixed a couple multi-touch scenarios that were drawing erratic lines [MOD-638]
+
+v2.0	Upgraded to module api version 2 for 1.8.0.1
 
 v1.1	Added multi-touch support [MOD-243]
 		Added "image" property to the paint view. See example and documentation to find out more.
